refactor(obs): move OBS-websocket polling from tripbot to vlc-server (#564)
* refactor(obs): move PollStreamingActive goroutine from tripbot to vlc-server

OBS polling no longer runs in the tripbot cmd. The goroutine moves to
cmd/vlc-server alongside the libvlc stats poller; vlc-server is the
natural home for OBS-facing integration since it already lives on OBS's
data plane (dashcam RTSP, onscreens HTTP). Keeps tripbot decoupled from
OBS — tripbot is a chat/scoreboard service and doesn't otherwise need
to know OBS exists.


* refactor(instrumentation): relocate OBS gauges to vlc_server.go

OBS gauges (obs_streaming_active, obs_active_fps, obs_stream_output_*,
and the OBSStats / OBSStreaming interfaces) move from tripbot.go to
vlc_server.go since the polling now runs in the vlc-server cmd. Same
package, no callsite changes; the file split makes ownership obvious.


---------
