+++ cmd/tripbot/tripbot.go
@@ -18,7 +18,6 @@ import (
 	terrors "github.com/adanalife/tripbot/pkg/errors"
 	"github.com/adanalife/tripbot/pkg/helpers"
 	onscreensClient "github.com/adanalife/tripbot/pkg/onscreens-client"
-	"github.com/adanalife/tripbot/pkg/obs"
 	"github.com/adanalife/tripbot/pkg/server"
 	"github.com/adanalife/tripbot/pkg/telemetry"
 	mytwitch "github.com/adanalife/tripbot/pkg/twitch"
@@ -73,7 +72,6 @@ func main() {
 	findInitialVideo()
 	users.InitLeaderboard(context.Background())
 	startCron()
-	startOBSPolling()
 	loadTwitchToken()   // must precede chatbot.Initialize — provides the IRC token
 	setUpTwitchClient() // required for the below
 	updateSubscribers()
@@ -139,12 +137,6 @@ func startCron() {
 	// start cron and attach cronjobs
 	background.StartCron()
 	scheduleBackgroundJobs()
-}
-
-// startOBSPolling starts the background goroutine that polls OBS WebSocket
-// for streaming state and updates the obs_streaming_active gauge.
-func startOBSPolling() {
-	go obs.PollStreamingActive(context.Background(), 30*time.Second)
 }
 
 // loadTwitchToken pulls the bot's OAuth row from the oauth_tokens table.
