+++ pkg/instrumentation/tripbot.go
@@ -26,22 +26,6 @@ var (
 	twitchSubscribers = mustGauge("twitch_subscribers_total", "Current number of Twitch channel subscribers")
 	twitchFollowers   = mustGauge("twitch_followers_total", "Current number of Twitch channel followers")
 	twitchHelixErrors = mustCounter("twitch_helix_errors_total", "Total non-2xx responses from the Twitch Helix API, labeled by endpoint and status_code")
-	obsStreamingGauge = mustGauge("obs_streaming_active", "1 if OBS is actively streaming, 0 otherwise")
-
-	obsActiveFPS              = mustFloat64Gauge("obs_active_fps", "Current FPS being rendered by OBS")
-	obsAverageFrameRenderMS   = mustFloat64Gauge("obs_average_frame_render_time_ms", "Average time in milliseconds OBS spends rendering a frame")
-	obsCPUUsage               = mustFloat64Gauge("obs_cpu_usage_percent", "Current OBS CPU usage (percent)")
-	obsMemoryUsage            = mustFloat64Gauge("obs_memory_usage_mb", "Current OBS memory usage in MB")
-	obsRenderSkippedFrames    = mustFloat64Gauge("obs_render_skipped_frames", "Render-thread skipped frames since OBS started")
-	obsRenderTotalFrames      = mustFloat64Gauge("obs_render_total_frames", "Render-thread total frames since OBS started")
-	obsOutputSkippedFrames    = mustFloat64Gauge("obs_output_skipped_frames", "Output-thread skipped frames since OBS started")
-	obsOutputTotalFrames      = mustFloat64Gauge("obs_output_total_frames", "Output-thread total frames since OBS started")
-	obsStreamOutputBytes      = mustFloat64Gauge("obs_stream_output_bytes", "Bytes sent by the stream output (cumulative since stream start)")
-	obsStreamOutputDurationMS = mustFloat64Gauge("obs_stream_output_duration_ms", "Current stream output duration in milliseconds")
-	obsStreamCongestion       = mustFloat64Gauge("obs_stream_output_congestion", "Stream output congestion (0..1)")
-	obsStreamReconnecting     = mustGauge("obs_stream_output_reconnecting", "1 if the stream output is currently reconnecting, 0 otherwise")
-	obsStreamSkippedFrames    = mustFloat64Gauge("obs_stream_output_skipped_frames", "Stream-output skipped frames since stream start")
-	obsStreamTotalFrames      = mustFloat64Gauge("obs_stream_output_total_frames", "Stream-output total frames since stream start")
 )
 
 // ChatMessages exposes the chat-message counter through a tiny stable API
@@ -73,52 +57,6 @@ var TwitchAudience = twitchAudienceIface{subscribers: twitchSubscribers, followe
 // TwitchHelixErrors.Inc(endpoint, statusCode). Endpoint is a short label
 // like "GetUsers"; statusCode is the HTTP status reported by Twitch.
 var TwitchHelixErrors = twitchHelixErrorsIface{counter: twitchHelixErrors}
-
-// OBSStreaming exposes the streaming-active gauge.
-var OBSStreaming = obsStreamingIface{g: obsStreamingGauge}
-
-// OBSStatsSnapshot bundles OBS performance + stream-output stats so the
-// poller can publish in a single call without coupling instrumentation to
-// goobs types.
-type OBSStatsSnapshot struct {
-	ActiveFPS              float64
-	AverageFrameRenderTime float64 // ms
-	CPUUsage               float64 // percent
-	MemoryUsage            float64 // MB
-	RenderSkippedFrames    float64
-	RenderTotalFrames      float64
-	OutputSkippedFrames    float64
-	OutputTotalFrames      float64
-}
-
-// OBSStreamSnapshot is the stream-output side (only meaningful while
-// streaming, but always safe to publish — fields are zero when idle).
-type OBSStreamSnapshot struct {
-	OutputBytes       float64
-	OutputDurationMS  float64
-	OutputCongestion  float64
-	Reconnecting      bool
-	SkippedFrames     float64
-	TotalFrames       float64
-}
-
-// OBSStats exposes the OBS performance + stream-output gauges.
-var OBSStats = obsStatsIface{
-	activeFPS:           obsActiveFPS,
-	averageFrameRender:  obsAverageFrameRenderMS,
-	cpuUsage:            obsCPUUsage,
-	memoryUsage:         obsMemoryUsage,
-	renderSkippedFrames: obsRenderSkippedFrames,
-	renderTotalFrames:   obsRenderTotalFrames,
-	outputSkippedFrames: obsOutputSkippedFrames,
-	outputTotalFrames:   obsOutputTotalFrames,
-	streamBytes:         obsStreamOutputBytes,
-	streamDuration:      obsStreamOutputDurationMS,
-	streamCongestion:    obsStreamCongestion,
-	streamReconnecting:  obsStreamReconnecting,
-	streamSkipped:       obsStreamSkippedFrames,
-	streamTotal:         obsStreamTotalFrames,
-}
 
 type chatCounterIface struct{ counter metric.Int64Counter }
 
@@ -170,59 +108,6 @@ func (h twitchHelixErrorsIface) Inc(endpoint string, statusCode int) {
 		attribute.String("endpoint", endpoint),
 		attribute.Int("status_code", statusCode),
 	))
-}
-
-type obsStreamingIface struct{ g metric.Int64Gauge }
-
-func (o obsStreamingIface) Set(active bool) {
-	v := int64(0)
-	if active {
-		v = 1
-	}
-	o.g.Record(context.Background(), v)
-}
-
-type obsStatsIface struct {
-	activeFPS           metric.Float64Gauge
-	averageFrameRender  metric.Float64Gauge
-	cpuUsage            metric.Float64Gauge
-	memoryUsage         metric.Float64Gauge
-	renderSkippedFrames metric.Float64Gauge
-	renderTotalFrames   metric.Float64Gauge
-	outputSkippedFrames metric.Float64Gauge
-	outputTotalFrames   metric.Float64Gauge
-	streamBytes         metric.Float64Gauge
-	streamDuration      metric.Float64Gauge
-	streamCongestion    metric.Float64Gauge
-	streamReconnecting  metric.Int64Gauge
-	streamSkipped       metric.Float64Gauge
-	streamTotal         metric.Float64Gauge
-}
-
-func (o obsStatsIface) Update(s OBSStatsSnapshot) {
-	ctx := context.Background()
-	o.activeFPS.Record(ctx, s.ActiveFPS)
-	o.averageFrameRender.Record(ctx, s.AverageFrameRenderTime)
-	o.cpuUsage.Record(ctx, s.CPUUsage)
-	o.memoryUsage.Record(ctx, s.MemoryUsage)
-	o.renderSkippedFrames.Record(ctx, s.RenderSkippedFrames)
-	o.renderTotalFrames.Record(ctx, s.RenderTotalFrames)
-	o.outputSkippedFrames.Record(ctx, s.OutputSkippedFrames)
-	o.outputTotalFrames.Record(ctx, s.OutputTotalFrames)
-}
-
-func (o obsStatsIface) UpdateStream(s OBSStreamSnapshot) {
-	ctx := context.Background()
-	o.streamBytes.Record(ctx, s.OutputBytes)
-	o.streamDuration.Record(ctx, s.OutputDurationMS)
-	o.streamCongestion.Record(ctx, s.OutputCongestion)
-	v := int64(0)
-	if s.Reconnecting {
-		v = 1
-	}
-	o.streamReconnecting.Record(ctx, v)
-	o.streamSkipped.Record(ctx, s.SkippedFrames)
-	o.streamTotal.Record(ctx, s.TotalFrames)
 }
 
 func mustCounter(name, desc string) metric.Int64Counter {
